+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org).
 
+## [Unreleased]
+
+### Added
+
+- wsusserver: support nesting of computer target groups
+
 ## [v1.1.2](https://github.com/tragiccode/tragiccode-wsusserver/tree/v1.1.2) (2019-09-17)
 
 [Full Changelog](https://github.com/tragiccode/tragiccode-wsusserver/compare/v1.1.1...v1.1.2)